requirements.txt
# Flask 核心
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8

# 数据库
SQLAlchemy==2.0.40

# 表单验证
WTForms==3.2.1
email_validator==2.2.0
dnspython==2.7.0

# 安全/加密
cryptography==44.0.2

# HTTP 请求
requests==2.32.3
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

# 环境变量
python-dotenv==1.1.0

# 其他
blinker==1.9.0
cffi==1.17.1
pycparser==2.22
MarkupSafe==3.0.2
typing_extensions==4.13.1

# WSGI 服务器
waitress==3.0.2

# ============================================
# 阿里云 DNS SDK (可选)
# 如果不使用阿里云 DNS，可以注释掉以下依赖
# ============================================
alibabacloud_alidns20150109==4.3.0
alibabacloud-tea-openapi==0.4.3
alibabacloud-tea-util==0.3.14
alibabacloud-credentials==1.0.5
